requirements.txt
tiktoken
pypdf
openai
fastapi
python-dotenv
numpy
yt-dlp
requests
youtube-transcript-api
python-dotenv
langchain
langchain-huggingface
langchain-openrouter
supabase
argon2-cffi
jinja2
langchain-core
pinecone
pytest
pytest-cov
httpx
